Convenience methods for issuing a single HTTP request.
Part of the "instant gratification" is for a new Snice Testing developer
to just be able to create a test in no time, execute it and see the
result. This means that a lot of boilerplate is being generated by
default (e.g. you don't have to specify an HTTP Protocol and if you
don't, then we'll create a default one for you with sane values).

None of this should really be part of any production deployments since
then you'll probably want more control over e.g. the HTTP stack settings
(timeouts etc).

See the example test case Http101