Adding a fromString to all SRIs to accomondate default Deserialization support from Jetty
In Jetty (and I assume it is som JAX-RS whatever specification thing),
when deserializing incoming webtraffic, such as a form param or a path
param, from a string into a java object, it will by default look for
a static method called fromString. If it exists, it'll use it. If not,
you would have to wrap that into a "param converter" instead.

This seemed simpler and doesn't really break the non-dependency on
irrelevant frameworks for these simple POJOs (I wouldn't have done it
if I actually had to be dependent on some jax-rs stuff).